requirements.txt
accelerate==0.29.3
datasets==2.18.0
huggingface-hub==0.21.4
joblib==1.3.2
jupyter_client 
jupyter_core 
jupyterlab_widgets==3.0.10
librosa==0.10.1
lightgbm==4.3.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
parquet==1.3.1
Pillow==10.0.1
polars==0.20.16
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.12.0
stop-words==2018.7.23
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm 
transformers==4.39.1
